requirements.txt
mlflow
pandas
scikit-learn
onnx
skl2onnx
python-dotenv
